+++ CHANGELOG.md
@@ -65,6 +65,8 @@ BUG FIXES:
     merged for the proper before/after chain. [GH-1555]
   - Use the Vagrant temporary directory again for temporary files
     since they can be quite large and were messing with tmpfs. [GH-1442]
+  - Fix issue parsing extra SSH args in `vagrant ssh` in multi-machine
+    environments. [GH-1545]
 
 ## 1.1.6 (April 3, 2013)